requirements.txt
anthropic>=0.18.0
pydantic>=2.0.0
rich>=13.0.0
python-dotenv>=1.0.0
asyncio-pool>=0.6.0
